refactor: adjust implementation of `DiscussionHero`
